nouveauyama-main/backend/requirements.txt
# GROUPE YAMA+ Backend Dependencies
# Compatible avec VPS autonome (sans packages propriétaires)

# Core Framework
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2

# Database
motor==3.3.1
pymongo==4.5.0

# Authentication & Security
bcrypt==4.1.3
PyJWT==2.10.1
python-jose==3.5.0
passlib==1.7.4
cryptography==46.0.3

# HTTP & Async
aiohttp==3.13.2
httpx==0.28.1
requests==2.32.5

# Validation & Serialization
pydantic==2.12.5
email-validator==2.3.0
python-multipart==0.0.20

# Environment & Configuration
python-dotenv==1.2.1

# AI Integration (OpenAI direct - pas de packages propriétaires)
openai==1.99.9

# Email Services
mailersend==2.0.0

# Push Notifications
pywebpush==2.2.0
py-vapid==1.9.4

# PDF Generation
reportlab==4.4.7
PyPDF2==3.0.1

# Image Processing
pillow==12.0.0

# Scheduling
APScheduler==3.11.2

# Payments
stripe==14.1.0

# Utils
python-dateutil==2.9.0.post0
pytz==2025.2
